Move function axiosPostForm to LoginUtils.tsx
